backend/requirements.txt
# FastAPI Backend Dependencies
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
boto3==1.34.34
websockets==12.0
aioboto3==13.0.1
aiofiles==23.2.1
python-dotenv==1.0.1
redis==5.0.1
celery==5.3.6
